+++ packages/common/utils/getButtonSpacing.ts
@@ -27,7 +27,6 @@ const getCacheKey = ({
   return `${compact}-${flush}-${hasFrontier}-${startIcon}-${endIcon}`;
 };
 
-/** @deprecated will be removed in v6.0.0 Please use getButtonSpacingProps moving forward */
 export const getButtonSpacing = memoize(function getButtonSpacing({
   compact,
   flush,