+++ packages/common/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## Unreleased
+
+#### 📘 Misc
+
+- Undeprecated getButtonSpacing.
+
 ## 5.31.3 ((1/30/2024, 02:05 PM PST))
 
 This is an artificial version bump with no new change.